refactor: create generic form fields and versions specific to workflows with variables
